MM-433, MM-476: Corrected vitalsigns-bloodglucose slicing definition and replaced LOINC code for fasting blood glucose from 41604-0 to 14770-2.
BloodGlucose derives from Laboratory-Observation. This base profiles slices based on system, but BloodGlucose tries to slice on both system and code. That is not legal. To remedy, the slicing was reset to the base definition and the explicit inline profile codes have been externalized to 2 new ValueSets. The net effect is the same, but it doesn't trip validation anymore.